Reorganize BLT commands into Uiowa namespace
