+++ composer.json
@@ -29,12 +29,12 @@
     },
     "autoload": {
         "psr-4": {
-            "Sitenow\\": "blt/src/"
+            "Uiowa\\": "blt/src/"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "Sitenow\\Tests\\PHPUnit\\": "tests/phpunit/src/"
+            "Uiowa\\Tests\\PHPUnit\\": "tests/phpunit/src/"
         }
     },
     "config": {
